+++ src/components/common/FormGroupInlineControlLabel.js
